finished writing selinger opitimizer except TODO comment points...
