+++ planner/optimizer/selinger_optimizer.go
@@ -1,13 +1,19 @@
 package optimizer
 
 import (
+	"fmt"
+	mapset "github.com/deckarep/golang-set/v2"
 	stack "github.com/golang-collections/collections/stack"
 	pair "github.com/notEpsilon/go-pair"
 	"github.com/ryogrid/SamehadaDB/catalog"
+	"github.com/ryogrid/SamehadaDB/common"
+	"github.com/ryogrid/SamehadaDB/execution/executors"
 	"github.com/ryogrid/SamehadaDB/execution/expression"
 	"github.com/ryogrid/SamehadaDB/execution/plans"
 	"github.com/ryogrid/SamehadaDB/parser"
 	"github.com/ryogrid/SamehadaDB/samehada/samehada_util"
+	"github.com/ryogrid/SamehadaDB/storage/access"
+	"github.com/ryogrid/SamehadaDB/storage/table/column"
 	"github.com/ryogrid/SamehadaDB/storage/table/schema"
 	"github.com/ryogrid/SamehadaDB/types"
 	"math"
@@ -70,6 +76,23 @@ type SelingerOptimizer struct {
 func NewSelingerOptimizer() *SelingerOptimizer {
 	// TODO: (SDB) not implemented yet
 	return nil
+}
+
+func containsAny(map1 mapset.Set[string], map2 mapset.Set[string]) bool {
+	interSet := map1.Intersect(map2)
+	if interSet.Cardinality() > 0 {
+		return true
+	} else {
+		return false
+	}
+}
+
+func makeSet[T comparable](from []*T) mapset.Set[T] {
+	joined := mapset.NewSet[T]()
+	for _, f := range from {
+		joined.Add(*f)
+	}
+	return joined
 }
 
 func isColumnName(v interface{}) bool {
@@ -150,15 +173,12 @@ func (so *SelingerOptimizer) bestScan(selection []*parser.SelectFieldExpression,
 				colIdx := sc.GetColIndex(*exp.Left_.(*string))
 				if colIdx != math.MaxUint32 {
 					relatedOps = append(relatedOps, exp)
-					/*
-							auto iter = ranges.find(offset);
-							if (iter != ranges.end()) {
-								iter->second.Update(be.Op(),
-									be.Right()->AsConstantValue().GetValue(),
-									Range::Dir::kRight);
-							}
-						}
-					*/
+					//auto iter = ranges.find(offset);
+					//if (iter != ranges.end()) {
+					//	iter->second.Update(be.Op(),
+					//		be.Right()->AsConstantValue().GetValue(),
+					//		Range::Dir::kRight);
+					//}
 					if rng, ok := ranges[int(colIdx)]; ok {
 						rng.Update(exp.ComparisonOperationType_, exp.Right_.(*types.Value), DIR_RIGHT)
 					}
@@ -170,15 +190,12 @@ func (so *SelingerOptimizer) bestScan(selection []*parser.SelectFieldExpression,
 				colIdx := sc.GetColIndex(*exp.Right_.(*string))
 				if colIdx != math.MaxUint32 {
 					relatedOps = append(relatedOps, exp)
-					/*
-							auto iter = ranges.find(offset);
-							if (iter != ranges.end()) {
-								iter->second.Update(be.Op(),
-									be.Left()->AsConstantValue().GetValue(),
-									Range::Dir::kLeft);
-							}
-						}
-					*/
+					//auto iter = ranges.find(offset);
+					//if (iter != ranges.end()) {
+					//	iter->second.Update(be.Op(),
+					//		be.Left()->AsConstantValue().GetValue(),
+					//		Range::Dir::kLeft);
+					//}
 					if rng, ok := ranges[int(colIdx)]; ok {
 						rng.Update(exp.ComparisonOperationType_, exp.Left_.(*types.Value), DIR_LEFT)
 					}
@@ -193,10 +210,10 @@ func (so *SelingerOptimizer) bestScan(selection []*parser.SelectFieldExpression,
 	if len(relatedOps) > 0 {
 		scanExp = samehada_util.ConvParsedBinaryOpExprToExpIFOne(relatedOps[0])
 		for ii := 1; ii < len(relatedOps); ii++ {
-			/*
-			   scan_exp =
-			       BinaryExpressionExp(scan_exp, BinaryOperation::kAnd, related_ops[i]);
-			*/
+			// append addiitional condition with AND operator
+
+			//scan_exp = BinaryExpressionExp(scan_exp, BinaryOperation::kAnd, related_ops[i]);
+			scanExp = expression.AppendLogicalCondition(scanExp, expression.AND, samehada_util.ConvParsedBinaryOpExprToExpIFOne(relatedOps[ii]))
 		}
 	}
 
@@ -210,17 +227,15 @@ func (so *SelingerOptimizer) bestScan(selection []*parser.SelectFieldExpression,
 		}
 		//targetIndex := from.GetIndex(candidates[key])
 		// Plan new_plan = IndexScanSelect(from, target_idx, stat, *span.min,*span.max, scan_exp, select);
-		newPlan := plans.NewRangeScanWithIndexPlanNode(schema.ConvSelectFieldExpToOutputSchema(selection), from.OID(), int32(key), scanExp, span.min, span.max)
+		var newPlan plans.Plan = plans.NewRangeScanWithIndexPlanNode(schema.ConvSelectFieldExpToOutputSchema(selection), from.OID(), int32(key), nil, span.min, span.max)
 		// if (!TouchOnly(scan_exp, from.GetSchema().GetColumn(key).Name())) {
 		if !touchOnly(scanExp, sc.GetColumn(uint32(key)).GetColumnName()) {
-			/*
-			   new_plan = std::make_shared<SelectionPlan>(new_plan, scan_exp, stat);
-			*/
+			//new_plan = std::make_shared<SelectionPlan>(new_plan, scan_exp, stat);
+			newPlan = plans.NewSelectionPlanNode(newPlan, newPlan.OutputSchema(), scanExp)
 		}
 		if len(selection) != int(newPlan.OutputSchema().GetColumnCount()) {
-			/*
-			   new_plan = std::make_shared<ProjectionPlan>(new_plan, select);
-			*/
+			//new_plan = std::make_shared<ProjectionPlan>(new_plan, select);
+			newPlan = plans.NewProjectionPlanNode(newPlan, samehada_util.ConvParsedSelectionExprToExpIFOne(selection))
 		}
 		if newPlan.AccessRowCount() < minimamCost {
 			bestScan = newPlan
@@ -228,21 +243,16 @@ func (so *SelingerOptimizer) bestScan(selection []*parser.SelectFieldExpression,
 		}
 	}
 
-	/*
-	  Plan full_scan_plan(new FullScanPlan(from, stat));
-	*/
-	fullScanPlan := plans.NewSeqScanPlanNode(sc, scanExp, from.OID())
+	// Plan full_scan_plan(new FullScanPlan(from, stat));
+	fullScanPlan := plans.NewSeqScanPlanNode(sc, nil, from.OID())
 	if scanExp != nil {
-		/*
-		   full_scan_plan =
-		       std::make_shared<SelectionPlan>(full_scan_plan, scan_exp, stat);
-		*/
+		// full_scan_plan = std::make_shared<SelectionPlan>(full_scan_plan, scan_exp, stat);
+		fullScanPlan = plans.NewSeqScanPlanNode(sc, scanExp, from.OID())
 	}
 
 	if len(selection) != int(sc.GetColumnCount()) {
-		/*
-			full_scan_plan = std::make_shared<ProjectionPlan>(full_scan_plan, select);
-		*/
+		// full_scan_plan = std::make_shared<ProjectionPlan>(full_scan_plan, select);
+		fullScanPlan = plans.NewProjectionPlanNode(fullScanPlan, samehada_util.ConvParsedSelectionExprToExpIFOne(selection))
 	}
 	if fullScanPlan.AccessRowCount() < minimamCost {
 		bestScan = fullScanPlan
@@ -250,6 +260,37 @@ func (so *SelingerOptimizer) bestScan(selection []*parser.SelectFieldExpression,
 	}
 
 	return bestScan, nil
+}
+
+func findBestScans(query *parser.QueryInfo, exec_ctx *executors.ExecutorContext, c *catalog.Catalog, txn *access.Transaction) map[mapset.Set[string]]CostAndPlan {
+	// TODO: (SDB) not implemented yet (findBestScans)
+
+	optimalPlans := make(map[mapset.Set[string]]CostAndPlan)
+
+	// 1. Initialize every single tables to start.
+	touchedColumns := query.WhereExpression_.TouchedColumns()
+	for _, item := range query.SelectFields_ {
+		touchedColumns = touchedColumns.Union(item.TouchedColumns())
+	}
+	for _, from := range query.JoinTables_ {
+		tbl := c.GetTableByName(*from)
+		stats := c.GetTableByName(*from).GetStatistics()
+		projectTarget := make([]*column.Column, 0)
+		// Push down all selection & projection.
+		for ii := 0; ii < int(tbl.GetColumnNum()); ii++ {
+			for _, touchedCol := range touchedColumns.ToSlice() {
+				tableCol := tbl.Schema().GetColumn(uint32(ii))
+				// TODO: (SDB) GetColumnName() value should contain table name. if not,  rewrite of this comparison code is needed
+				if tableCol.GetColumnName() == touchedCol.GetColumnName() {
+					projectTarget = append(projectTarget, tableCol)
+				}
+			}
+		}
+		scan, _ := NewSelingerOptimizer().bestScan(query.SelectFields_, query.WhereExpression_, tbl, c, stats)
+		optimalPlans[makeSet([]*string{from})] = CostAndPlan{scan.AccessRowCount(), scan}
+	}
+
+	return optimalPlans
 }
 
 // TODO: (SDB) caller should pass *where* args which is deep copied
@@ -362,12 +403,51 @@ func (so *SelingerOptimizer) bestJoin(where *parser.BinaryOpExpression, left pla
 	return candidates[0], nil
 }
 
+func findBestJoin(optimalPlans map[mapset.Set[string]]CostAndPlan, query *parser.QueryInfo, exec_ctx *executors.ExecutorContext, c *catalog.Catalog, txn *access.Transaction) plans.Plan {
+	// TODO: (SDB) need to setup of optimalPlans which is needed at BestJoin
+	// TODO: (SDB) need to setup of statistics data of tables related to query
+
+	for ii := 0; ii < len(query.JoinTables_); ii += 1 {
+		for baseTableFrom, baseTableCP := range optimalPlans {
+			for joinTableFrom, joinTableCP := range optimalPlans {
+				if containsAny(baseTableFrom, joinTableFrom) {
+					continue
+				}
+				// TODO: (SDB) (len(baseTable) + len(joinTable) == ii + 1) should be checked?
+				//             and (len(baseTable) == 1 or len(joinTable) == 1) should be checked?
+				bestJoinPlan, _ := NewSelingerOptimizer().bestJoin(query.WhereExpression_, baseTableCP.plan, joinTableCP.plan)
+				fmt.Println(bestJoinPlan)
+
+				joinedTables := baseTableFrom.Union(joinTableFrom)
+				common.SH_Assert(1 < joinedTables.Cardinality(), "joinedTables.Cardinality() is illegal!")
+				cost := bestJoinPlan.AccessRowCount()
+
+				// TODO: (SDB) update target should be changed to tempolal table?
+				//             (its scope is same ii value loop and it is merged to optimalPlans at end of the ii loop)
+				if existedPlan, ok := optimalPlans[joinedTables]; ok {
+					optimalPlans[joinedTables] = CostAndPlan{cost, bestJoinPlan}
+				} else if cost < existedPlan.cost {
+					optimalPlans[joinedTables] = CostAndPlan{cost, bestJoinPlan}
+				}
+			}
+		}
+	}
+	optimalPlan, ok := optimalPlans[makeSet(query.JoinTables_)]
+	samehada_util.SHAssert(ok, "plan which includes all tables is not found")
+
+	// Attach final projection and emit the result
+	solution := optimalPlan.plan
+	solution = plans.NewProjectionPlanNode(solution, samehada_util.ConvParsedSelectionExprToExpIFOne(query.SelectFields_))
+
+	return solution
+}
+
 // TODO: (SDB) caller should check predicate whether it is optimizable and
 //	           if not, caller can't call this function
 //             (predicate including bracket or OR operation case is not supported now)
 
 // TODO: (SDB) adding support of ON clause (Optimize, bestJoin, bestScan)
-func (so *SelingerOptimizer) Optimize() (plans.Plan, error) {
-	// TODO: (SDB) not implemented yet
-	return nil, nil
+func (so *SelingerOptimizer) Optimize(query *parser.QueryInfo, exec_ctx *executors.ExecutorContext, c *catalog.Catalog, txn *access.Transaction) (plans.Plan, error) {
+	optimalPlans := findBestScans(query, exec_ctx, c, txn)
+	return findBestJoin(optimalPlans, query, exec_ctx, c, txn), nil
 }
